Move BaseAudioContext to separate file
